fix(chat): fix name parsing, fix file unshare modal opening
